requirements/common.txt
asgiref==3.2.7
Django==3.0.6
pytz==2020.1
sqlparse==0.3.1
Pillow==7.1.1
psycopg2==2.8.5
django-ckeditor==5.9.0
django-bleach==0.6.1
django-money==1.1
django-filter==2.2.0
django-crispy-forms==1.9.1
django-honeypot==0.8.0
django-redis==4.11.0
celery==4.4.2
requests==2.23.0
geoip2==3.0.0
django-ipware==2.1.0
django-solo==1.1.3
pyyaml==5.3.1
squareup==5.2.2.20200422
django-localflavor==3.0.1
cryptography==2.9.2
